Add unused linter
