+++ src/cipher/secp256k1-go/secp256_test.go
@@ -9,8 +9,10 @@ import (
 	"testing"
 )
 
-const TESTS = 1    //10000 // how many tests
-const SigSize = 65 //64+1
+const TESTS = 10000 // how many tests
+
+// nolint: unused
+const SigSize = 65 // 64+1
 
 func Test_Secp256_00(t *testing.T) {
