+++ src/cipher/secp256k1-go/secp256k1-go2/xyz_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 )
 
-func _TestGejDouble(t *testing.T) { //nolint: deadcode
+func _TestGejDouble(t *testing.T) { //nolint: deadcode,unused
 	var a, aExp, r XYZ
 	a.X.SetHex("79be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798")
 	a.Y.SetHex("483ada7726a3c4655da4fbfc0e1108a8fd17b448a68554199c47d08ffb10d4b8")
