+++ src/visor/db.go
@@ -86,7 +86,7 @@ func CheckDatabase(db *dbutil.DB, pubkey cipher.PubKey, quit chan struct{}) erro
 }
 
 // backup the corrypted db first, then rebuild the history DB.
-func rebuildHistoryDB(db *dbutil.DB, history *historydb.HistoryDB, bc *Blockchain, quit chan struct{}) (*dbutil.DB, error) {
+func rebuildHistoryDB(db *dbutil.DB, history *historydb.HistoryDB, bc *Blockchain, quit chan struct{}) (*dbutil.DB, error) { // nolint: unused
 	db, err := backupDB(db)
 	if err != nil {
 		return nil, err
@@ -133,7 +133,7 @@ func rebuildHistoryDB(db *dbutil.DB, history *historydb.HistoryDB, bc *Blockchai
 }
 
 // backupDB makes a backup copy of the DB
-func backupDB(db *dbutil.DB) (*dbutil.DB, error) {
+func backupDB(db *dbutil.DB) (*dbutil.DB, error) { // nolint: unused
 	// backup the corrupted database
 	dbReadOnly := db.IsReadOnly()
 
@@ -172,7 +172,7 @@ func RepairCorruptDB(db *dbutil.DB, pubkey cipher.PubKey, quit chan struct{}) (*
 	}
 }
 
-func rebuildCorruptDB(db *dbutil.DB, pubkey cipher.PubKey, quit chan struct{}) (*dbutil.DB, error) { //nolint: deadcode
+func rebuildCorruptDB(db *dbutil.DB, pubkey cipher.PubKey, quit chan struct{}) (*dbutil.DB, error) { //nolint: deadcode,unused
 	history := historydb.New()
 	bc, err := NewBlockchain(db, BlockchainConfig{Pubkey: pubkey})
 	if err != nil {
@@ -230,7 +230,7 @@ func moveCorruptDB(dbPath string) (string, error) {
 }
 
 // copyCorruptDB copy a file to makeCorruptDBPath(dbPath)
-func copyCorruptDB(dbPath string) (string, error) {
+func copyCorruptDB(dbPath string) (string, error) { // nolint: unused
 	newDBPath, err := makeCorruptDBPath(dbPath)
 	if err != nil {
 		return "", err
